Relocate title iteration to its own module
Bug: https://phabricator.wikimedia.org/T140266
